+++ glue/viewers/table/qt/viewer_widget.py
@@ -63,7 +63,11 @@ class DataTableModel(QtCore.QAbstractTableModel):
             return None
 
         if orientation == Qt.Horizontal:
-            return self.columns[section].label
+            column_name = self.columns[section].label
+            units = self._data.get_component(self.columns[section]).units
+            if units != '':
+                column_name += " [{0}]".format(units)
+            return column_name
         elif orientation == Qt.Vertical:
             return str(self.order[section])